Exercise 5: Positioning - Not OK
A commitment just for the record.
The solution was based on an article [ https://coder-coder.com/display-divs-side-by-side/ ] (Add space between divs by using a margin, and it will still fit!)
By just adding margin-right to the left ‘box’ the result is for the left box to be wider than the right one.
Instead I will apply width: 45% for the .article items in the .flexible container and hopefully there will be a 10% gap between the two boxes.